Edge device files/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
deepface==0.0.93
dnspython==2.7.0
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==20181003210633
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.13.0
huggingface_hub==0.29.2
HyperPyYAML==1.2.2
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
onnx==1.17.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.1.0
piwho==1.3.2
protobuf==5.29.3
PyAudio==0.2.14
pycparser==2.22
Pygments==2.19.1
pymongo==4.11.2
pyparsing==3.2.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
smbus2==0.5.0
sounddevice==0.5.1
soupsieve==2.6
speechbrain==1.0.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
tflite-runtime==2.14.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
